gather: Fix removal of empty file from gather
The MG currently uses `-empty` with find to remove
the empty files. This only works for files with a
size of 0 bytes.

In cases where the file has an empty line or a single
character this doesn't work.

This patch updates the logic to also consider files with
LF endings (just a new line)

Signed-off-by: Niraj Yadav <niryadav@redhat.com>